Refactorizando un poquito para quitar Pandas como dependencia y adoptar el nuevo formato en columna.
